Fix main's pc status not displayed if the actual server port and nat port are different
The API ping redirect to equavalent image, but it will return the server port.
If NAT port is different, the image can't be loaded.
So it now runs on client side. A javascript func will get ping2 api to check PC status.
Then another func to assign the image according to the status.

Added blank-32.png as the default status icon.